backend/requirements.txt
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2>=2.9
Pillow>=10.3
django-cors-headers>=4.3
